Added new menu for Quickplay.
The new menu is in gGame.cpp.
In gServerBrowser there is the function that is called when the menu item is selected, which takes care of communicating with the server to get a match for the player.
In english_base.txt there are the new definitions for the names used in the menu (recreate configure file to see the changes)